made a refactoring. final
