+++ data/initial_file.txt
@@ -1,5 +1,5 @@
-Загадочный мир нанотехнологий…
+Загадочный мир нанотехнологий… 
 
 Приставкой «нано» обозначают миллиардную долю метра. Приблизительно такой размер имеют молекулы, из-за чего иногда нанотехнологии называют молекулярными технологиями.
 
-Молекулярные технологии представляют собой общность методов, которые позволяют манипулировать атомами и молекулами. Другими словами, с их помощью можно создавать и эксплуатировать материалы, строить определённые системы, в основе которых находятся наноструктуры с размерами компонентов от 1 до 100 нанометров
+Молекулярные технологии представляют собой общность методов, которые позволяют манипулировать атомами и молекулами. Другими словами, с их помощью можно создавать и эксплуатировать материалы, строить определённые системы, в основе которых находятся наноструктуры с размерами компонентов от 1 до 100 нанометров.
